Release 1.22.0

+++ Lib9c/Module/Guild/GuildParticipantModule.cs
@@ -1,12 +1,10 @@
 #nullable enable
 using System;
 using System.Diagnostics.CodeAnalysis;
-using System.Numerics;
 using Bencodex.Types;
 using Lib9c;
-using Libplanet.Types.Assets;
-using Nekoyume.Action.Guild.Migration.LegacyModels;
 using Nekoyume.Model.Guild;
+using Nekoyume.Module.ValidatorDelegation;
 using Nekoyume.TypedAddress;
 using Nekoyume.ValidatorDelegation;
 
@@ -28,25 +26,34 @@ namespace Nekoyume.Module.Guild
             GuildAddress guildAddress,
             AgentAddress target)
         {
+            var height = repository.ActionContext.BlockIndex;
+            var signer = repository.ActionContext.Signer;
             if (repository.TryGetGuildParticipant(target, out _))
             {
-                throw new ArgumentException("The signer already joined a guild.");
+                throw new InvalidOperationException("The signer already joined a guild.");
             }
 
-            if (repository.GetGuildRejoinCooldown(target) is { } cooldown
-                && cooldown.Cooldown(repository.ActionContext.BlockIndex) > 0L)
+            if (repository.GetDelegator(target) is { } delegator
+                && delegator.UnbondingRefs.Count > 0)
             {
                 throw new InvalidOperationException(
-                    $"The signer is in the rejoin cooldown period until block {cooldown.ReleaseHeight}");
+                    $"The signer cannot join guild while unbonding");
+            }
+
+            var validatorRepository = new ValidatorRepository(repository.World, repository.ActionContext);
+            if (validatorRepository.TryGetDelegatee(signer, out var _))
+            {
+                throw new InvalidOperationException("Validator cannot join a guild.");
             }
 
             var guildParticipant = new GuildParticipant(target, guildAddress, repository);
             var guildGold = repository.GetBalance(guildParticipant.DelegationPoolAddress, Currencies.GuildGold);
+            var guild = repository.GetGuild(guildAddress);
             repository.SetGuildParticipant(guildParticipant);
             repository.IncreaseGuildMemberCount(guildAddress);
             if (guildGold.RawValue > 0)
             {
-                repository.Delegate(target, guildGold);
+                guildParticipant.Delegate(guild, guildGold, height);
             }
 
             return repository;
@@ -57,26 +64,33 @@ namespace Nekoyume.Module.Guild
             AgentAddress guildParticipantAddress,
             GuildAddress dstGuildAddress)
         {
+            var height = repository.ActionContext.BlockIndex;
             var guildParticipant1 = repository.GetGuildParticipant(guildParticipantAddress);
             var srcGuild = repository.GetGuild(guildParticipant1.GuildAddress);
             var dstGuild = repository.GetGuild(dstGuildAddress);
+            if (srcGuild.Address == dstGuild.Address)
+            {
+                throw new InvalidOperationException("The signer cannot move to the same guild.");
+            }
+
+            var srcGuildDelegatee = repository.GetDelegatee(srcGuild.ValidatorAddress);
             var validatorRepository = new ValidatorRepository(repository.World, repository.ActionContext);
-            var srcValidatorDelegatee = validatorRepository.GetValidatorDelegatee(srcGuild.ValidatorAddress);
-            var dstValidatorDelegatee = validatorRepository.GetValidatorDelegatee(dstGuild.ValidatorAddress);
+            var dstValidatorDelegatee = validatorRepository.GetDelegatee(dstGuild.ValidatorAddress);
             if (dstValidatorDelegatee.Tombstoned)
             {
                 throw new InvalidOperationException("The validator of the guild to move to has been tombstoned.");
             }
 
             var guildParticipant2 = new GuildParticipant(guildParticipantAddress, dstGuildAddress, repository);
-            var bond = validatorRepository.GetBond(srcValidatorDelegatee, guildParticipantAddress);
+            var bond = repository.GetBond(srcGuildDelegatee, guildParticipantAddress);
+            var share = bond.Share;
             repository.RemoveGuildParticipant(guildParticipantAddress);
             repository.DecreaseGuildMemberCount(guildParticipant1.GuildAddress);
             repository.SetGuildParticipant(guildParticipant2);
             repository.IncreaseGuildMemberCount(dstGuildAddress);
-            if (bond.Share > 0)
+            if (share > 0)
             {
-                repository.Redelegate(guildParticipantAddress, dstGuildAddress);
+                guildParticipant1.Redelegate(srcGuild, dstGuild, share, height);
             }
 
             return repository;
@@ -103,18 +117,25 @@ namespace Nekoyume.Module.Guild
                     "The signer is a guild master. Guild master cannot quit the guild.");
             }
 
-            var validatorRepository = new ValidatorRepository(repository.World, repository.ActionContext);
-            var validatorDelegatee = validatorRepository.GetValidatorDelegatee(guild.ValidatorAddress);
-            var bond = validatorRepository.GetBond(validatorDelegatee, agentAddress);
+            if (repository.GetDelegator(agentAddress).UnbondingRefs.Count > 0)
+            {
+                throw new InvalidOperationException(
+                    $"The signer cannot leave guild while unbonding");
+            }
+
+            var height = repository.ActionContext.BlockIndex;
+            var guildParticipant = repository.GetGuildParticipant(agentAddress);
+            var delegatee = repository.GetDelegatee(guild.ValidatorAddress);
+            var bond = repository.GetBond(delegatee, agentAddress);
+            var share = bond.Share;
+
+            if (bond.Share > 0)
+            {
+                guildParticipant.Undelegate(guild, share, height);
+            }
 
             repository.RemoveGuildParticipant(agentAddress);
             repository.DecreaseGuildMemberCount(guild.Address);
-            if (bond.Share > 0)
-            {
-                repository.Undelegate(agentAddress);
-            }
-
-            repository.SetGuildRejoinCooldown(agentAddress, repository.ActionContext.BlockIndex);
 
             return repository;
         }
@@ -135,96 +156,5 @@ namespace Nekoyume.Module.Guild
                 return false;
             }
         }
-
-        // TODO: Hide this method as private after migration.
-        public static GuildRepository Delegate(
-            this GuildRepository repository,
-            AgentAddress guildParticipantAddress,
-            FungibleAssetValue fav)
-        {
-            var height = repository.ActionContext.BlockIndex;
-
-            // TODO: Remove below unnecessary height condition after migration.
-            if (repository.World.GetDelegationMigrationHeight() is long migrationHeight)
-            {
-                height = Math.Max(height, migrationHeight);
-            }
-
-            var guildParticipant = repository.GetGuildParticipant(guildParticipantAddress);
-            var guild = repository.GetGuild(guildParticipant.GuildAddress);
-            guildParticipant.Delegate(guild, fav, height);
-
-            return repository;
-        }
-
-        private static GuildRepository Undelegate(
-            this GuildRepository repository,
-            AgentAddress guildParticipantAddress)
-        {
-            var height = repository.ActionContext.BlockIndex;
-            var guildParticipant = repository.GetGuildParticipant(guildParticipantAddress);
-            var guild = repository.GetGuild(guildParticipant.GuildAddress);
-            var share = repository.GetBond(
-                new ValidatorRepository(repository.World, repository.ActionContext)
-                    .GetValidatorDelegatee(guild.ValidatorAddress),
-                guildParticipantAddress).Share;
-            guildParticipant.Undelegate(guild, share, height);
-
-            return repository;
-        }
-
-        private static GuildRepository Undelegate(
-            this GuildRepository repository,
-            AgentAddress guildParticipantAddress,
-            BigInteger share)
-        {
-            var height = repository.ActionContext.BlockIndex;
-            var guildParticipant = repository.GetGuildParticipant(guildParticipantAddress);
-            var guild = repository.GetGuild(guildParticipant.GuildAddress);
-            guildParticipant.Undelegate(guild, share, height);
-
-            return repository;
-        }
-
-        public static GuildRepository Redelegate(
-            this GuildRepository repository,
-            AgentAddress guildParticipantAddress,
-            GuildAddress dstGuildAddress)
-        {
-            var height = repository.ActionContext.BlockIndex;
-            var guildParticipant = repository.GetGuildParticipant(guildParticipantAddress);
-            var guild = repository.GetGuild(guildParticipant.GuildAddress);
-            var share = repository.GetBond(
-                new ValidatorRepository(repository.World, repository.ActionContext)
-                    .GetValidatorDelegatee(guild.ValidatorAddress),
-                guildParticipantAddress).Share;
-            var dstGuild = repository.GetGuild(dstGuildAddress);
-            guildParticipant.Redelegate(guild, dstGuild, share, height);
-
-            return repository;
-        }
-
-        private static GuildRepository SetGuildRejoinCooldown(
-            this GuildRepository repository,
-            AgentAddress guildParticipantAddress,
-            long height)
-        {
-            var guildRejoinCooldown = new GuildRejoinCooldown(guildParticipantAddress, height);
-            repository.UpdateWorld(
-                repository.World.SetAccount(
-                    Addresses.GuildRejoinCooldown,
-                    repository.World.GetAccount(Addresses.GuildRejoinCooldown)
-                        .SetState(guildParticipantAddress, guildRejoinCooldown.Bencoded)));
-            return repository;
-        }
-
-        private static GuildRejoinCooldown? GetGuildRejoinCooldown(
-            this GuildRepository repository,
-            AgentAddress guildParticipantAddress)
-            => repository.World
-                .GetAccount(Addresses.GuildRejoinCooldown)
-                .GetState(guildParticipantAddress) is Integer bencoded
-                    ? new GuildRejoinCooldown(guildParticipantAddress, bencoded)
-                    : null;
     }
 }
